requirements.txt
mlflow>=2.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
matplotlib>=3.7.0
pyyaml>=6.0.0
joblib>=1.3.0
# Optionnel
xgboost>=2.0.0
imbalanced-learn>=0.12.0
